+++ gatsby-node.mjs
@@ -1,0 +1,11 @@
+import { createSchemaCustomization } from './gatsby/createSchemaCustomization.mjs';
+import { createPages } from './gatsby/createPages.mjs';
+import { createResolvers } from './gatsby/createResolvers.mjs';
+import { onCreateNode } from './gatsby/onCreateNode.mjs';
+
+export {
+  createSchemaCustomization,
+  createPages,
+  createResolvers,
+  onCreateNode
+};
