+++ package.json
@@ -28,7 +28,7 @@
     "clipboardy": "^3.0.0",
     "copy-to-clipboard": "^3.3.3",
     "gatsby": "^5.7.0",
-    "gatsby-awesome-pagination": "^0.3.8",
+    "gatsby-core-utils": "^4.7.0",
     "gatsby-plugin-image": "^3.7.0",
     "gatsby-plugin-mdx": "^5.7.0",
     "gatsby-plugin-netlify": "^5.1.0",
